Rename tsconfig.libs to tsconfig.spec
