+++ libs/tools/export/vault-export/vault-export-core/jest.config.js
@@ -1,6 +1,6 @@
 const { pathsToModuleNameMapper } = require("ts-jest");
 
-const { compilerOptions } = require("../../../../shared/tsconfig.libs");
+const { compilerOptions } = require("../../../../shared/tsconfig.spec");
 
 /** @type {import('jest').Config} */
 module.exports = {
